+++ CHANGELOG.md
@@ -23,6 +23,10 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
+
+## [1.3.2] - 2025-11-12
+
+* Add more helpful error message for AKS min/max knot selection
 * Add support for holdout set in `GoodnessOfFitCheck`.
 * Introduce modules needed for Meridian Scenario Planner and add
 `scenarioplanner` extra.
@@ -415,4 +419,5 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 [1.2.1]: https://github.com/google/meridian/releases/tag/v1.2.1
 [1.3.0]: https://github.com/google/meridian/releases/tag/v1.3.0
 [1.3.1]: https://github.com/google/meridian/releases/tag/v1.3.1
-[Unreleased]: https://github.com/google/meridian/compare/v1.3.1...HEAD
+[1.3.2]: https://github.com/google/meridian/releases/tag/v1.3.2
+[Unreleased]: https://github.com/google/meridian/compare/v1.3.2...HEAD